+++ src/app/layout.tsx
@@ -2,6 +2,8 @@ import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
 import { Toaster } from "sonner";
+import flightPlansData from "@/data/flightPlans.json" assert { type: "json" };
+import Sidebar from "@/components/sidebar";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -28,7 +30,11 @@ export default function RootLayout({
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
-        {children}
+        <div className="grid grid-cols-[250px_1fr] h-screen">
+          <Sidebar plans={flightPlansData} />
+
+          {children}
+        </div>
         <Toaster richColors />
       </body>
     </html>
