+++ src/components/flightPlanItem.tsx
@@ -1,24 +1,22 @@
 import { FlightPlan } from "@/interfaces/flightPlan";
 import "@/styles/flightPlanItem.css";
 import clsx from "clsx";
+import Link from "next/link";
 
 interface FlightPlanProps {
   plan: FlightPlan;
   selected: boolean;
-  onClick: () => void;
 }
 
-export default function FlightPlanItem({
-  plan,
-  selected,
-  onClick,
-}: FlightPlanProps) {
+export default function FlightPlanItem({ plan, selected }: FlightPlanProps) {
   return (
-    <li
-      className={clsx("fp-list-item", { "fp-list-item-selected": selected })}
-      onClick={onClick}
-    >
-      {plan.dep} - {plan.dest}
+    <li>
+      <Link
+        className={clsx("fp-list-item", { "fp-list-item-selected": selected })}
+        href={`/fp/${plan.id.toString()}`}
+      >
+        {plan.dep} - {plan.dest}
+      </Link>
     </li>
   );
 }
